mv download func to ppocr/utils/network.py
